+++ lib/dice/diceTypes.go
@@ -1,7 +1,7 @@
 package dice
 
 import (
-	"internal/dice"
+	"dice/roll"
 )
 
 var (
@@ -11,7 +11,7 @@ var (
 // D2 rolls a D2 (1-2)
 func D2(count int) int {
 	for count > 0 {
-		result += dice.Roll(2)
+		result += roll.Roll(2)
 		count--
 	}
 	return result
@@ -20,7 +20,7 @@ func D2(count int) int {
 // D4 rolls a D4 (1-4)
 func D4(count int) int {
 	for count > 0 {
-		result += dice.Roll(4)
+		result += roll.Roll(4)
 		count--
 	}
 	return result
@@ -29,7 +29,7 @@ func D4(count int) int {
 // D6 rolls a D6 (1-6)
 func D6(count int) int {
 	for count > 0 {
-		result += dice.Roll(6)
+		result += roll.Roll(6)
 		count--
 	}
 	return result
@@ -38,7 +38,7 @@ func D6(count int) int {
 // D8 rolls a D8 (1-8)
 func D8(count int) int {
 	for count > 0 {
-		result += dice.Roll(8)
+		result += roll.Roll(8)
 		count--
 	}
 	return result
@@ -47,7 +47,7 @@ func D8(count int) int {
 // D10 rolls a D10 (1-10)
 func D10(count int) int {
 	for count > 0 {
-		result += dice.Roll(10)
+		result += roll.Roll(10)
 		count--
 	}
 	return result
@@ -56,7 +56,7 @@ func D10(count int) int {
 // D12 rolls a D12 (1-12)
 func D12(count int) int {
 	for count > 0 {
-		result += dice.Roll(12)
+		result += roll.Roll(12)
 		count--
 	}
 	return result
@@ -65,7 +65,7 @@ func D12(count int) int {
 // D20 rolls a D20 (1-20)
 func D20(count int) int {
 	for count > 0 {
-		result += dice.Roll(20)
+		result += roll.Roll(20)
 		count--
 	}
 	return result
@@ -74,20 +74,24 @@ func D20(count int) int {
 // Can only roll one at a time for the moment
 func D00(count int) int {
 	for count > 0 {
-		a := dice.Roll(10)
+		// Roll a D10
+		a := roll.Roll(10)
 
+		// Change the D10 into a D00
 		if a == 10 {
 			a = 0
 		} else {
 			a = a * 10
 		}
 
-		b := dice.Roll(10)
+		// Roll a D10
+		b := roll.Roll(10)
 
 		if b == 1 && a == 0 {
 			return 100
 		}
 
+		// Add the results together
 		result = a + b
 		count--
 	}
