--- lib/dice/internal/go.mod
@@ -1,3 +1,0 @@
-module internal/dice
-
-go 1.20
