+++ lib/dice/roll/go.mod
@@ -1,0 +1,3 @@
+module dice/roll
+
+go 1.20
